wip commit to merge latest swagger2.0 branch
