+++ bravado/client.py
@@ -51,24 +51,20 @@ without caching
 
 """
 
-from bravado.compat import json
 import logging
 import time
 import urllib
 import urlparse
 
-from yelp_uri import urllib_utf8
-
-import swagger_type
+from bravado.compat import json
 from bravado.http_client import APP_JSON, SynchronousHttpClient
 from bravado.mapping.model import build_models
-from bravado.response import HTTPFuture, post_receive
+from bravado.mapping.resource import build_resources
 from bravado.swagger_model import (
     Loader,
     is_file_scheme_uri,
 )
 from bravado.swagger_type import SwaggerTypeCheck
-
 
 log = logging.getLogger(__name__)
 
@@ -158,58 +154,6 @@ def get_client(*args, **kwargs):
     return cache(*args, **kwargs)
 
 
-class Operation(object):
-    """Perform a request by taking the kwargs passed to the call and
-    constructing an HTTP request.
-    """
-
-    def __init__(self, uri, operation, http_client, models):
-        self._uri = uri
-        self._json = operation
-        self._http_client = http_client
-        self._models = models
-        self.__doc__ = create_operation_docstring(operation)
-
-    def __repr__(self):
-        return u"%s(%s)" % (self.__class__.__name__, self._json[u'nickname'])
-
-    def _construct_request(self, **kwargs):
-        _request_options = kwargs.pop('_request_options', {}) or {}
-
-        request = {}
-        request['method'] = self._json[u'method']
-        request['url'] = self._uri
-        request['params'] = {}
-        request['headers'] = _request_options.get('headers', {}) or {}
-
-        for param in self._json.get(u'parameters', []):
-            value = kwargs.pop(param[u'name'], param.get('defaultValue'))
-            validate_and_add_params_to_request(param, value, request,
-                                               self._models)
-        if kwargs:
-            raise TypeError(u"'%s' does not have parameters %r" % (
-                self._json[u'nickname'], kwargs.keys()))
-        return request
-
-    def __call__(self, **kwargs):
-        log.debug(u"%s?%r" % (
-            self._json[u'nickname'],
-            urllib_utf8.urlencode(kwargs)))
-        request = self._construct_request(**kwargs)
-
-        def response_future(response, **kwargs):
-            # Assume status is OK, an exception would have been raised already
-            if not response.text:
-                return None
-
-            return post_receive(
-                response.json(),
-                swagger_type.get_swagger_type(self._json),
-                self._models,
-                **kwargs)
-        return HTTPFuture(self._http_client, request, response_future)
-
-
 def get_resource_url(base_path, url_base, resource_base_path):
     if base_path:
         return base_path
@@ -221,102 +165,6 @@ def get_resource_url(base_path, url_base, resource_base_path):
         return url_base.rstrip('/') + resource_base_path
 
     return resource_base_path
-
-
-class Resource(object):
-    """Swagger resource, described in an API declaration.
-    """
-
-    def __init__(self, name, operations):
-        log.debug(u"Building resource '%s'" % name)
-        self._name = name
-        self._operations = operations
-
-    # XXX 1.2
-    # @classmethod
-    # def from_api_doc(cls, api_doc, http_client, base_path, url_base=None):
-    #     """
-    #     :param api_doc: api doc which defines this resource
-    #     :type  api_doc: :class:`dict`
-    #     :param http_client: a :class:`bravado.http_client.HttpClient`
-    #     :param base_path: base url to perform api requests. Used to override
-    #             the path provided in the api spec
-    #     :param url_base: a url used as the base for resource definitions
-    #             that include a relative basePath
-    #     """
-    #     declaration = api_doc['api_declaration']
-    #     models = build_models(declaration.get('models', {}))
-    #
-    #     def build_operation(api_obj, operation):
-    #         log.debug(u"Building operation %s.%s" % (
-    #             api_obj.get('name'), operation['nickname']))
-    #
-    #         resource_base_path = declaration.get('basePath')
-    #         url = get_resource_url(base_path, url_base, resource_base_path)
-    #         url = url.rstrip('/') + api_obj['path']
-    #         return Operation(url, operation, http_client, models)
-    #
-    #     operations = dict(
-    #         (oper['nickname'], build_operation(api, oper))
-    #         for api in declaration['apis']
-    #         for oper in api['operations'])
-    #     return cls(api_doc['name'], operations)
-
-    @classmethod
-    def from_path(cls, path_name, path_dict, api_url, http_client):
-        """
-        :param path_name: Path of the resource. ex: /pets, pets/{id},
-        :param path_dict: json-like dict which defines the resource. The key
-            is usually an http method (get, put, post, delete, options, head,
-            patch)
-        :param api_url: base URL used to service API requests
-        :param http_client: a :class:`bravado.http_client.HttpClient`
-        """
-        # TODO: Fix for 2.0
-
-        # XXX 1.2
-        # declaration = api_doc['api_declaration']
-        # models = build_models(declaration.get('models', {}))
-
-        # TODO: path_name can be a non-http method: 'parameters'
-        # TODO: path_name can be $ref
-
-        def build_operation(http_method, operation_dict):
-            log.debug(u"Building operation %s.%s" % (
-                path_name, operation_dict['operationId']))
-
-            # resource_base_path = declaration.get('basePath')
-            # url = get_resource_url(base_path, url_base, resource_base_path)
-            # url = url.rstrip('/') + api_obj['path']
-
-            url = api_url + path_name
-
-            # TODO: figure out where to get 'models' from
-            return Operation(url, operation_dict, http_client, models=None)
-
-        operations = {}
-        for http_method, operation_dict in path_dict.items():
-            operation = build_operation(http_method, operation_dict)
-            operations[operation['nickname']] = operation
-
-        return cls(path_name, operations)
-
-    def __repr__(self):
-        return u"%s(%s)" % (self.__class__.__name__, self._name)
-
-    def __getattr__(self, item):
-        """
-        :param item: name of the :class:`Operation` to return
-        :return: an :class:`Operation`
-        """
-        op = self._operations.get(item)
-        if not op:
-            raise AttributeError(u"Resource '%s' has no operation '%s'" %
-                                 (self._name, item))
-        return op
-
-    def __dir__(self):
-        return self._operations.keys()
 
 
 class SwaggerClient(object):
@@ -416,11 +264,10 @@ class SwaggerClient(object):
         definitions = build_models(spec['definitions'])
         spec['x_definitions'] = definitions
 
-        # TODO: Pick up with 'resources' here
-        # resources = build_resources(spec, http_client)
-        resources = None
+        # START branch port_resource_and_ops
+        resources = build_resources(spec, http_client)
+        # END branch port_resource_and_ops
 
-        # STOP HERE
         return cls(spec['x_api_url'], resources)
 
     def __repr__(self):
@@ -501,89 +348,6 @@ def build_api_serving_url(spec, origin_url, preferred_scheme=None):
     scheme = pick_a_scheme(spec.get('schemes'))
 
     return urlparse.urlunparse((scheme, netloc, path, None, None, None))
-
-
-def build_resources(spec, http_client):
-    """Transforms the REST resources in the json-like spec into rich :Resource:
-    objects that have associated :Operation:s.
-
-    :param spec: json-like spec in dict form
-    :param http_client: an HTTP client used to perform requests
-    :type  http_client: :class:`bravado.http_client.HttpClient`
-    :returns: dict where (key,value) = (resource name, Resource instance)
-    """
-    resources = {}
-    paths = spec['paths']
-    for path_name, path_dict in paths:
-        resources[path_name] = Resource.from_path(
-            path_name,
-            path_dict,
-            spec['x_api_url'],
-            http_client)
-
-    return resources
-
-
-def _build_param_docstring(param):
-    """Builds param docstring from the param dict
-
-    :param param: data to create docstring from
-    :type param: dict
-    :returns: string giving meta info
-
-    Example: ::
-        status (string) : Statuses to be considered for filter
-        from_date (string) : Start date filter"
-    """
-    string = "\t" + param.get("name")
-    type_ = param.get('$ref') or param.get('format') or param.get('type')
-    if type_:
-        string += (" (%s) " % type_)
-    if param.get('description'):
-        string += ": " + param["description"]
-    return string + "\n"
-
-
-def create_operation_docstring(json_):
-    """Builds Operation docstring from the json dict
-
-    :param json_: data to create docstring from
-    :type json_: dict
-    :returns: string giving meta info
-
-    Example: ::
-
-        client.pet.findPetsByStatus?
-
-    Outputs: ::
-
-        [GET] Finds Pets by status
-
-        Multiple status values can be provided with comma seperated strings
-        Args:
-                status (string) : Statuses to be considered for filter
-                from_date (string) : Start date filter
-        Returns:
-                array
-        Raises:
-                400: Invalid status value
-    """
-    docstring = ""
-    if json_.get('summary'):
-        docstring += ("[%s] %s\n\n" % (json_['method'], json_.get('summary')))
-    docstring += (json_["notes"] + "\n") if json_.get("notes") else ''
-
-    if json_["parameters"]:
-        docstring += "Args:\n"
-        for param in json_["parameters"]:
-            docstring += _build_param_docstring(param)
-    if json_.get('type'):
-        docstring += "Returns:\n\t%s\n" % json_["type"]
-    if json_.get('responseMessages'):
-        docstring += "Raises:\n"
-        for msg in json_.get('responseMessages'):
-            docstring += "\t%s: %s\n" % (msg.get("code"), msg.get("message"))
-    return docstring
 
 
 def handle_form_param(name, value, type_, request):
