+++ tests/mapping/resource/build_resources_test.py
@@ -1,0 +1,9 @@
+from mock import Mock
+
+from bravado.mapping.resource import build_resources
+from bravado.http_client import HttpClient
+
+
+def test_empty():
+    resources = build_resources({}, http_client=Mock(spec=HttpClient))
+    print resources
