+++ tests/mapping/resource/conftest.py
@@ -1,0 +1,118 @@
+import pytest
+
+from bravado.mapping.model import create_model_type
+
+
+@pytest.fixture
+def paths_dict():
+    # The '#/paths' dict from a spec that define the REST resources
+    return {
+        "User": {
+            "properties": {
+                "id": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "username": {
+                    "type": "string"
+                },
+                "firstName": {
+                    "type": "string"
+                },
+                "lastName": {
+                    "type": "string"
+                },
+                "email": {
+                    "type": "string"
+                },
+                "password": {
+                    "type": "string"
+                },
+                "phone": {
+                    "type": "string"
+                },
+                "userStatus": {
+                    "type": "integer",
+                    "format": "int32",
+                    "description": "User Status"
+                }
+            },
+            "xml": {
+                "name": "User"
+            }
+        },
+        "Category": {
+            "properties": {
+                "id": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "name": {
+                    "type": "string"
+                }
+            },
+        },
+        "Pet": {
+            "required": [
+                "name",
+                "photoUrls"
+            ],
+            "properties": {
+                "id": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "category": {
+                    "$ref": "#/definitions/Category"
+                },
+                "name": {
+                    "type": "string",
+                    "example": "doggie"
+                },
+                "photoUrls": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                },
+                "tags": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/Tag"
+                    }
+                },
+            },
+        },
+        "Tag": {
+            "properties": {
+                "id": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "name": {
+                    "type": "string"
+                }
+            },
+            "xml": {
+                "name": "Tag"
+            }
+        },
+    }
+
+
+@pytest.fixture
+def user_model(definitions_dict):
+    user_dict = definitions_dict['User']
+    return create_model_type('User', user_dict)
+
+
+@pytest.fixture
+def tag_model(definitions_dict):
+    tag_dict = definitions_dict['Tag']
+    return create_model_type('Tag', tag_dict)
+
+
+@pytest.fixture
+def pet_dict(definitions_dict):
+    return definitions_dict['Pet']
+
